Add prometheus deleting registry function in main conftest
